backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pydantic
alembic
PyJWT==2.11.0
pwdlib[argon2]==0.2.1
python-dotenv
python-multipart
Pillow
pillow-heif
psycopg2-binary
gunicorn
supabase
pytest
pytest-asyncio
requests
ruff
